# Conflicts:
#	client/socket-client/app/src/main/java/com/microsoft/pacmania/MapsActivity.java

+++ client/socket-client/app/src/main/java/com/microsoft/pacmania/Fruit.java
@@ -1,11 +1,11 @@
 package com.microsoft.pacmania;
 
 public class Fruit {
-    public double value;
+    public int value;
     public String name;
     public LocationData location;
 
-    public Fruit(double value, String name, LocationData location) {
+    public Fruit(int value, String name, LocationData location) {
         this.value = value;
         this.name = name;
         this.location = location;
